# PostgreSQL: Optimise table count metrics