--- a/postgres/changelog.d/17109.changed
+++ b/postgres/changelog.d/17109.changed
@@ -1,0 +1,1 @@
+PostgreSQL: Optimise table count query