--- a/postgres/tests/common.py
+++ b/postgres/tests/common.py
@@ -206,11 +206,8 @@ def check_common_metrics(aggregator, expected_tags, count=1):
 
 def check_db_count(aggregator, expected_tags, count=1):
     table_count = 7
-    # We create 2 additional partition tables when partition is available
+    # We create 2 additional partition tables when partition is available + 2 parent tables
     if float(POSTGRES_VERSION) >= 11.0:
-        table_count = 9
-    # And PG >= 14 will also report the parent table
-    if float(POSTGRES_VERSION) >= 14.0:
         table_count = 11
     aggregator.assert_metric(
         'postgresql.table.count',